% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcPR}
\alias{calcPR}
\title{calcPR}
\usage{
calcPR(X, B, by = 5 * 1e-04)
}
\arguments{
\item{X}{estimated matrix of GRN or eQTL-mapping}

\item{B}{ground true value in simulation}

\item{by}{precision of curve}
}
\value{
list of PRcurve
}
\description{
calculated precision-recall curve with specified quantile cutoff
}
