% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{cv.ridgeRegression2}
\alias{cv.ridgeRegression2}
\title{cv.ridgeRegression2}
\usage{
cv.ridgeRegression2(X, Y, Sk, ngamma = 100, nfold = 5, n, p, k)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{Gene expression matrix}

\item{Sk}{eQTL index of genes}

\item{ngamma}{number of hyper-parameter in CV}

\item{nfold}{CVfold number. Default 5/10}

\item{n}{number of observations}

\item{p}{number of genes}

\item{k}{number of eQTLs}
}
\value{
gamma_min optimal gamma to minimize cross-validation error
}
\description{
cv.ridgeRegression2
}
\examples{
gamma = cv.ridgeRegression2(data$Data$X, data$Data$Y, data$Data$Sk, ngamma = 20, nfold = 5, 100, 10, 60)
}
