% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadGTExv6p}
\alias{downloadGTExv6p}
\title{downloadGTEx}
\usage{
downloadGTExv6p(type = "genes", file = NULL, ...)
}
\value{
ExpressionSet set
}
\description{
Download GTeX borrow from Yarn
}
\examples{
# obj <- downloadGTExv6p(type = 'genes', file = '/media/xinchou/Storage/gtexv6p.rds')
}
