% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{eQTLdataExt}
\alias{eQTLdataExt}
\title{eQTLdataExt}
\usage{
eQTLdataExt(B, F, Sk, QTL, n, k, sigma2 = 0.01, intercept = 5)
}
\arguments{
\item{B}{Gene regulatory network matrix}

\item{F}{eQTL mapping matrix}

\item{sigma2}{noise of observations}
}
\description{
generate simulated network structure for eQTLSEM algorithm, with fixed GRN and eQTL mapping
require("eQTLsem")
}
