% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{G}
\alias{G}
\title{Genotype of Human Sapiens from GTEx}
\format{5 candidate genotype pool extracted from GTEx patients}
\usage{
G
}
\description{
A dataset containing Genotype of SNPs for different genes
}
\keyword{datasets}
