% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{implSSEM}
\alias{implSSEM}
\title{implSSEM}
\usage{
implSSEM(data = NULL)
}
\arguments{
\item{data}{Data archive of experiment measurements, includeing
eQTL matrices, Gene expression matrices of different conditions,
marker of eQTLs and data generation SEM model}
}
\value{
List of TPR and FDR
}
\description{
implementor function of SSEM solver for eQTL mapping
}
