% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{initLambdaSSEM2B}
\alias{initLambdaSSEM2B}
\title{initLambdaSSEM}
\usage{
initLambdaSSEM2B(X, Y, fit, Wb)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{Gene expression matrix}

\item{fit}{SEM fit Object as lambda = Inf}

\item{Wb}{weight matrix of B for adaptive lasso terms. Default as 1 / B(init) from eidge regression.}
}
\description{
Initialized Lambda coefficients (lambda, rho) for Sparse SEM
}
