% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{initLambdaSSEM2F}
\alias{initLambdaSSEM2F}
\title{initLambdaSSEM}
\usage{
initLambdaSSEM2F(X, Y, Sk, Wb, Wf, p, q)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{Gene expression matrix}

\item{Sk}{eQTL index of genes}

\item{Wb}{weight matrix of B for adaptive lasso terms. Default as 1 / B(init) from eidge regression.}

\item{Wf}{weight matrix of F for adaptive lasso terms. Default as 1 / F(init) from ridge regression.}

\item{p}{number of genes}

\item{q}{number of eQTLs}
}
\description{
Initialized Lambda coefficients (lambda, rho) for Sparse SEM
}
