% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{initRegression}
\alias{initRegression}
\title{initRegression}
\usage{
initRegression(X, Y, Sk, n, p, k)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{Gene expression matrix}

\item{Sk}{eQTL index of genes}

\item{n}{number of observations}

\item{p}{number of genes}

\item{k}{number of eQTLs}
}
\description{
initRegression
}
