% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lamax.ridgeRegression}
\alias{lamax.ridgeRegression}
\title{lamax.ridgeRegression}
\usage{
lamax.ridgeRegression(X, Y, Sk, n, p, k)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{gene expression matrix}
}
\value{
max lambda
}
\description{
lamax.ridgeRegression
}
