% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obj.ridgeRegression}
\alias{obj.ridgeRegression}
\title{obj.ridgeRegression}
\usage{
obj.ridgeRegression(X, Y, fit, trans = F)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{gene expression matrix}

\item{fit}{regression fit result object}

\item{trans}{if rows for sample, trans = TRUE, otherwise, trans = FALSE. Default FALSE}
}
\value{
error squared norm of ||(I-B)Y - FX||_2^2
}
\description{
obj.ridgeRegression
}
