% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{opt.SSEMiPALM}
\alias{opt.SSEMiPALM}
\title{opt.SSEMiPALM}
\usage{
opt.SSEMiPALM(X, Y, B, F, Sk, sigma2, Wb = NULL, Wf = NULL,
  nlambda = 20, nrho = 20, p, wt = TRUE)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{Gene expression matrix}

\item{B}{initialized GRN-matrix}

\item{F}{initialized eQTL effect matrix}

\item{Sk}{eQTL index of genes}

\item{sigma2}{initialized noise variance from ridge regression}

\item{Wb}{weight matrix of B for adaptive lasso terms. Default as 1 / B(init) from eidge regression.}

\item{Wf}{weight matrix of F for adaptive lasso terms. Default as 1 / F(init) from ridge regression.}

\item{nlambda}{number of lambda}

\item{nrho}{number of rho}

\item{p}{number of genes}

\item{wt}{adaptive lasso or not. Default as TRUE}
}
\description{
Optimized Sparse SEM by BIC
}
