% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc.centerSSEM}
\alias{proc.centerSSEM}
\title{proc.centerSSEM}
\usage{
proc.centerSSEM(X, Y)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{gene expression matrix}
}
\value{
centered X and Y and mean vectors
}
\description{
proc.centerSSEM
}
