% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{randQtlsdata}
\alias{randQtlsdata}
\title{randQtlsdata}
\usage{
randQtlsdata(n, p, k, sparse = 1, s = 3, sigma2 = 0.01,
  intercept = 5, type = c("DG", "ER"), dag = TRUE, coefs = c(0.2,
  0.4), esize = c(0.5, 1), qdist = 100, ncore = 0.1)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of candidate genes}

\item{k}{number of candidate eQTLs (cis-eQTL + trans-eQTL)}

\item{sparse}{expected number of nonzero edges of each gene}

\item{s}{expected number of eQTLs}

\item{esize}{effect size of eQTL}
}
\description{
Generate simulated network structure for eQTLSEM algorithm, with defined
number of genes and number of cis-eQTLs and trans-eQTLs. To be different
with `randomeQTLdata`, `randQtlsdata`'s eQTL contains all Qtls with
require("eQTLsem")
}
