% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{ridgeRegression}
\alias{ridgeRegression}
\title{ridgeRegression}
\usage{
ridgeRegression(X, Y, Sk, gamma, n, p, k, trans = FALSE,
  sparse = FALSE)
}
\arguments{
\item{X}{eQTL matrix.}

\item{Y}{Gene expression matrix}

\item{Sk}{eQTL index of genes}

\item{gamma}{Hyperparameter for ridge regression}

\item{n}{number of observations}

\item{p}{number of genes}

\item{k}{number of eQTLs}

\item{trans}{if rows for sample, trans = TRUE, otherwise, trans = FALSE. Default FALSE}

\item{sparse}{Default is False}
}
\value{
fit List of SEM model
\describe{
\item{Bs}{ coefficient matrices of gene regulatory networks}
\item{fs}{ eQTL's coefficients w.r.t each gene}
\item{Fs}{ coefficient matrices of eQTL-gene effect}
\item{mu}{ Bias vector}
\item{sigma2}{ estimate of covariance in SEM}
}
}
\description{
Ridge regression for initialization
}
