% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rwiseGradient4SSEM2F}
\alias{rwiseGradient4SSEM2F}
\title{rwiseGradient4SSEM2F}
\usage{
rwiseGradient4SSEM2F(ImBi, Y, Xi, sigma2)
}
\arguments{
\item{ImBi}{updated (I - B)[i,]}

\item{Y}{Matrix of gene expression}

\item{Xi}{Submatrix of eQTLs}

\item{sigma2}{noise variance}
}
\value{
function whose argument is column vector fi
}
\description{
function generator function for network B
}
