% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{SSEMeNet}
\alias{SSEMeNet}
\title{SSEMeNet}
\usage{
SSEMeNet(
  X,
  Y,
  B,
  F,
  Sk,
  sigma2,
  lambda,
  alpha,
  rho,
  Wb = 1/abs(B),
  Wf = 1/abs(F),
  p,
  maxit = 100,
  inert = inert_opt("linear"),
  threshold = 1e-06,
  verbose = TRUE,
  sparse = TRUE,
  trans = FALSE,
  B2norm = NULL,
  strict = FALSE
)
}
\arguments{
\item{X}{eQTL matrix}

\item{Y}{Gene expression matrix}

\item{B}{initialized GRN-matrix}

\item{F}{initialized eQTL effect matrix}

\item{Sk}{eQTL index of genes}

\item{sigma2}{initialized noise variance from ridge regression}

\item{lambda}{Hyperparameter of lasso term in SSEM for B}

\item{alpha}{Hyperparameter for adaptive elastic net term in SSEM for F. Default (1 - alpha) x rho x ||F||_2^2 + alpha x rho x Wf x ||F||_1^1}

\item{rho}{Hyperparameter of adaptive elastic net lasso term in SSEM for F}

\item{Wb}{weight matrix of B for adaptive lasso terms. Default as 1 / B(init) from eidge regression.}

\item{Wf}{weight matrix of F for adaptive lasso terms. Default as 1 / F(init) from ridge regression.}

\item{p}{number of genes}

\item{maxit}{maximum iteration number. Default 100}

\item{inert}{inertial function for iPALM. Default as k-1/k+2}

\item{threshold}{convergence threshold. Default 1e-6}

\item{verbose}{Default TRUE}

\item{sparse}{Sparse Matrix or not}

\item{trans}{Fs matrix is transposed to k x p or not. If Fs from ridge regression, trans = TRUE, else, trans = FALSE}

\item{B2norm}{B2norm matrices upper bound generated from ridge regression. Default NULL.}

\item{strict}{Converge strictly or not. Default False}
}
\description{
Implementing iPALM algorithm to solve the Elstic-net based sparse SEM for eQTL mapping and network inference
}
\details{
Elstic-net based Sparse Structural Equation Model (SEM) with iPALM for eQTL mapping and GRN inference
}
