% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssemQr-pakcage.R
\docType{package}
\name{ssemQr}
\alias{ssemQr}
\alias{package-ssemQr}
\title{Solving Sparse Structural Equation Model}
\description{
Solving Sparse Structural Equation Model
}
\author{
Xin Zhou <\email{xxz220@miami.edu}>
}
